Fix issue creating new windows.
The `new-window` command takes a `-t` flag to indicate the target.
When there's a window with the same name as the session, ie:

    ocaml 1 ocaml

using `new-window -t ocaml -n vim` will raise a "index in use" error
because it's targeting the "ocaml" window (of the most recently used
session), not the "ocaml" session.

Using the `-a` flag tells tmux to append the window, so `-t` is read as
the session.